# src/generated/java/com/bci/users/api/CustomValidationApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.4.15).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package com.bci.users.api;

import com.bci.users.api.model.ApiCustomValidation;
import com.bci.users.api.model.ApiCustomValidationsResponse;
import com.bci.users.api.model.ApiErrorResponse;
import io.swagger.annotations.Api;
import io.swagger.annotations.ApiOperation;
import io.swagger.annotations.ApiParam;
import io.swagger.annotations.ApiResponse;
import io.swagger.annotations.ApiResponses;
import javax.validation.Valid;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;

@Api(value = "customValidation", description = "the customValidation API")
@RequestMapping(value = "")
public interface CustomValidationApi {

    @ApiOperation(value = "save or update custom validation", nickname = "customValidation", notes = "custom Validation", response = ApiCustomValidation.class, tags={ "customValidation", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "A successful status", response = ApiCustomValidation.class),
        @ApiResponse(code = 500, message = "Response with error.", response = ApiErrorResponse.class) })
    @RequestMapping(value = "/customValidation",
        produces = { "application/json" }, 
        consumes = { "application/json" },
        method = RequestMethod.POST)
    ResponseEntity<ApiCustomValidation> customValidation(@ApiParam(value = "custom validation."  )  @Valid @RequestBody ApiCustomValidation customValidation);


    @ApiOperation(value = "get all custom validations", nickname = "getCustomValidations", notes = "Returns all custom validations", response = ApiCustomValidationsResponse.class, tags={ "customvalidation", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "A list of users.", response = ApiCustomValidationsResponse.class),
        @ApiResponse(code = 500, message = "Response with error.", response = ApiErrorResponse.class) })
    @RequestMapping(value = "/customValidation",
        produces = { "application/json" }, 
        consumes = { "application/json" },
        method = RequestMethod.GET)
    ResponseEntity<ApiCustomValidationsResponse> getCustomValidations();

}