# src/generated/java/com/bci/users/api/LoginApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.4.15).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package com.bci.users.api;

import com.bci.users.api.model.ApiErrorResponse;
import com.bci.users.api.model.ApiUserLogin;
import com.bci.users.api.model.ApiUserResponse;
import io.swagger.annotations.Api;
import io.swagger.annotations.ApiOperation;
import io.swagger.annotations.ApiParam;
import io.swagger.annotations.ApiResponse;
import io.swagger.annotations.ApiResponses;
import javax.validation.Valid;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;

@Api(value = "login", description = "the login API")
@RequestMapping(value = "")
public interface LoginApi {

    @ApiOperation(value = "login", nickname = "userLogin", notes = "login", response = ApiUserResponse.class, tags={ "login", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "A successful status", response = ApiUserResponse.class),
        @ApiResponse(code = 500, message = "Response with error.", response = ApiErrorResponse.class) })
    @RequestMapping(value = "/login",
        produces = { "application/json" }, 
        consumes = { "application/json" },
        method = RequestMethod.POST)
    ResponseEntity<ApiUserResponse> userLogin(@ApiParam(value = "user login."  )  @Valid @RequestBody ApiUserLogin login);

}